subfs: Extract SubstitutingFS from scripttest package
This change extracts a util/subfs package to support an upcoming PR that makes
use of the same functionality outside of test code.
