+++ internal/sinktest/scripttest/scripttest.go
@@ -20,11 +20,11 @@ package scripttest
 
 import (
 	"embed"
-	"io"
 	"io/fs"
 	"strings"
 
 	"github.com/cockroachdb/cdc-sink/internal/util/ident"
+	"github.com/cockroachdb/cdc-sink/internal/util/subfs"
 )
 
 // ScriptFS contains the contents of the testdata directory.
@@ -35,48 +35,11 @@ var ScriptFS embed.FS
 // ScriptFSFor returns a wrapped version of ScriptFS that will
 // substitute the table and its schema into the returned files.
 func ScriptFSFor(tbl ident.Table) fs.FS {
-	return &SubstitutingFS{
+	return &subfs.SubstitutingFS{
 		FS: ScriptFS,
 		Replacer: strings.NewReplacer(
 			"{{ SCHEMA }}", tbl.Schema().Raw(),
 			"{{ TABLE }}", tbl.Raw(),
 		),
 	}
-}
-
-// SubstitutingFS performs string substitution on returned files.
-// This is used to replace sentinel values in the userscript with
-// dynamically generated values.
-type SubstitutingFS struct {
-	fs.FS
-	Replacer *strings.Replacer
-}
-
-// Open implements fs.FS.
-func (f *SubstitutingFS) Open(path string) (fs.File, error) {
-	file, err := f.FS.Open(path)
-	if err != nil {
-		return nil, err
-	}
-	info, err := file.Stat()
-	if err != nil {
-		return nil, err
-	}
-	buf, err := io.ReadAll(file)
-	if err != nil {
-		return nil, err
-	}
-	s := f.Replacer.Replace(string(buf))
-	return &substitutingFile{info, io.NopCloser(strings.NewReader(s))}, nil
-}
-
-type substitutingFile struct {
-	fs.FileInfo
-	io.ReadCloser
-}
-
-var _ fs.File = (*substitutingFile)(nil)
-
-func (s *substitutingFile) Stat() (fs.FileInfo, error) {
-	return s.FileInfo, nil
 }
